Refactor: move query function to queries subpackage
